fix generate graphql
