+++ web-app/graphql-codegen.ts
@@ -3,9 +3,9 @@ import type { CodegenConfig } from "@graphql-codegen/cli";
 const config: CodegenConfig = {
   overwrite: true,
   schema: "http://localhost:4000",
-  documents: "src/**/*.{ts,tsx}",
+  documents: ["src/gql/requests/mutations.ts", "src/gql/requests/queries.ts"],
   generates: {
-    "src/gql/": {
+    "src/gql/generate/": {
       preset: "client",
       plugins: [],
     },
