--- web-app/src/gql/graphql.ts
@@ -1,217 +1,0 @@
-/* eslint-disable */
-import { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/core';
-export type Maybe<T> = T | null;
-export type InputMaybe<T> = Maybe<T>;
-export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
-export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
-export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
-export type MakeEmpty<T extends { [key: string]: unknown }, K extends keyof T> = { [_ in K]?: never };
-export type Incremental<T> = T | { [P in keyof T]?: P extends ' $fragmentName' | '__typename' ? T[P] : never };
-/** All built-in and custom scalars, mapped to their actual values */
-export type Scalars = {
-  ID: { input: string; output: string; }
-  String: { input: string; output: string; }
-  Boolean: { input: boolean; output: boolean; }
-  Int: { input: number; output: number; }
-  Float: { input: number; output: number; }
-  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar.This scalar is serialized to a string in ISO 8601 format and parsed from a string in ISO 8601 format. */
-  DateTimeISO: { input: any; output: any; }
-  /** Geometry scalar type */
-  Geometry: { input: any; output: any; }
-};
-
-export type Category = {
-  __typename?: 'Category';
-  icon: Scalars['String']['output'];
-  id: Scalars['ID']['output'];
-  name: Scalars['String']['output'];
-};
-
-export type City = {
-  __typename?: 'City';
-  coordinates: Scalars['Geometry']['output'];
-  id: Scalars['ID']['output'];
-  name: Scalars['String']['output'];
-};
-
-export type Mutation = {
-  __typename?: 'Mutation';
-  addFavoritePlace: User;
-  createCategory: Category;
-  createPlace: Place;
-  deleteCategory: Category;
-  deletePlace: Place;
-  deleteUser: User;
-  removeFavoritePlace: User;
-  signIn: User;
-  signOut: User;
-  signUp: User;
-  updateCategory: Category;
-  updatePlace: Place;
-  updateUser: User;
-};
-
-
-export type MutationAddFavoritePlaceArgs = {
-  placeId: Scalars['String']['input'];
-};
-
-
-export type MutationCreateCategoryArgs = {
-  icon: Scalars['String']['input'];
-  name: Scalars['String']['input'];
-};
-
-
-export type MutationCreatePlaceArgs = {
-  address: Scalars['String']['input'];
-  categoryIds: Array<Scalars['String']['input']>;
-  city: Scalars['String']['input'];
-  coordinates: Scalars['Geometry']['input'];
-  description: Scalars['String']['input'];
-  name: Scalars['String']['input'];
-  ownerId?: InputMaybe<Scalars['String']['input']>;
-};
-
-
-export type MutationDeleteCategoryArgs = {
-  id: Scalars['ID']['input'];
-};
-
-
-export type MutationDeletePlaceArgs = {
-  id: Scalars['ID']['input'];
-};
-
-
-export type MutationDeleteUserArgs = {
-  id: Scalars['ID']['input'];
-};
-
-
-export type MutationRemoveFavoritePlaceArgs = {
-  placeId: Scalars['String']['input'];
-};
-
-
-export type MutationSignInArgs = {
-  email: Scalars['String']['input'];
-  password: Scalars['String']['input'];
-};
-
-
-export type MutationSignUpArgs = {
-  email: Scalars['String']['input'];
-  firstName: Scalars['String']['input'];
-  lastName: Scalars['String']['input'];
-  password: Scalars['String']['input'];
-};
-
-
-export type MutationUpdateCategoryArgs = {
-  icon: Scalars['String']['input'];
-  id: Scalars['ID']['input'];
-  name: Scalars['String']['input'];
-};
-
-
-export type MutationUpdatePlaceArgs = {
-  address: Scalars['String']['input'];
-  categoryIds?: InputMaybe<Array<Scalars['String']['input']>>;
-  city: Scalars['String']['input'];
-  coordinates: Scalars['Geometry']['input'];
-  description: Scalars['String']['input'];
-  id: Scalars['ID']['input'];
-  name: Scalars['String']['input'];
-};
-
-
-export type MutationUpdateUserArgs = {
-  email: Scalars['String']['input'];
-  firstName: Scalars['String']['input'];
-  id: Scalars['ID']['input'];
-  lastName: Scalars['String']['input'];
-  password: Scalars['String']['input'];
-};
-
-export type Place = {
-  __typename?: 'Place';
-  address: Scalars['String']['output'];
-  categories: Array<Category>;
-  city: City;
-  coordinates: Scalars['Geometry']['output'];
-  createdAt: Scalars['DateTimeISO']['output'];
-  description: Scalars['String']['output'];
-  id: Scalars['ID']['output'];
-  name: Scalars['String']['output'];
-  owner?: Maybe<User>;
-};
-
-export type Query = {
-  __typename?: 'Query';
-  categories: Array<Category>;
-  cities: Array<City>;
-  isAuthenticated: Scalars['Boolean']['output'];
-  isInFavorites: Scalars['Boolean']['output'];
-  myProfile: User;
-  place: Place;
-  places: Array<Place>;
-  users: Array<User>;
-};
-
-
-export type QueryIsInFavoritesArgs = {
-  placeId: Scalars['String']['input'];
-};
-
-
-export type QueryPlaceArgs = {
-  id: Scalars['ID']['input'];
-};
-
-
-export type QueryPlacesArgs = {
-  categoryIds?: InputMaybe<Array<Scalars['String']['input']>>;
-  city?: InputMaybe<Scalars['String']['input']>;
-};
-
-export type User = {
-  __typename?: 'User';
-  createdAt: Scalars['DateTimeISO']['output'];
-  email: Scalars['String']['output'];
-  favoritesPlaces: Array<Place>;
-  firstName: Scalars['String']['output'];
-  hashedPassword: Scalars['String']['output'];
-  id: Scalars['ID']['output'];
-  lastName: Scalars['String']['output'];
-  role: Scalars['String']['output'];
-  userInitials: Scalars['String']['output'];
-};
-
-export type SignInFormMutationVariables = Exact<{
-  email: Scalars['String']['input'];
-  password: Scalars['String']['input'];
-}>;
-
-
-export type SignInFormMutation = { __typename?: 'Mutation', signIn: { __typename?: 'User', id: string, email: string, firstName: string, lastName: string } };
-
-export type GetMyProfileSignInQueryVariables = Exact<{ [key: string]: never; }>;
-
-
-export type GetMyProfileSignInQuery = { __typename?: 'Query', myProfile: { __typename?: 'User', id: string, email: string, firstName: string, lastName: string } };
-
-export type SignUpMutationVariables = Exact<{
-  firstName: Scalars['String']['input'];
-  lastName: Scalars['String']['input'];
-  email: Scalars['String']['input'];
-  password: Scalars['String']['input'];
-}>;
-
-
-export type SignUpMutation = { __typename?: 'Mutation', signUp: { __typename?: 'User', id: string, firstName: string, lastName: string, email: string } };
-
-
-export const SignInFormDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"mutation","name":{"kind":"Name","value":"SignInForm"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"email"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"password"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"signIn"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"email"},"value":{"kind":"Variable","name":{"kind":"Name","value":"email"}}},{"kind":"Argument","name":{"kind":"Name","value":"password"},"value":{"kind":"Variable","name":{"kind":"Name","value":"password"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"email"}},{"kind":"Field","name":{"kind":"Name","value":"firstName"}},{"kind":"Field","name":{"kind":"Name","value":"lastName"}}]}}]}}]} as unknown as DocumentNode<SignInFormMutation, SignInFormMutationVariables>;
-export const GetMyProfileSignInDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"GetMyProfileSignIn"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"myProfile"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"email"}},{"kind":"Field","name":{"kind":"Name","value":"firstName"}},{"kind":"Field","name":{"kind":"Name","value":"lastName"}}]}}]}}]} as unknown as DocumentNode<GetMyProfileSignInQuery, GetMyProfileSignInQueryVariables>;
-export const SignUpDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"mutation","name":{"kind":"Name","value":"SignUp"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"firstName"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"lastName"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"email"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"password"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"String"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"signUp"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"firstName"},"value":{"kind":"Variable","name":{"kind":"Name","value":"firstName"}}},{"kind":"Argument","name":{"kind":"Name","value":"lastName"},"value":{"kind":"Variable","name":{"kind":"Name","value":"lastName"}}},{"kind":"Argument","name":{"kind":"Name","value":"email"},"value":{"kind":"Variable","name":{"kind":"Name","value":"email"}}},{"kind":"Argument","name":{"kind":"Name","value":"password"},"value":{"kind":"Variable","name":{"kind":"Name","value":"password"}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"id"}},{"kind":"Field","name":{"kind":"Name","value":"firstName"}},{"kind":"Field","name":{"kind":"Name","value":"lastName"}},{"kind":"Field","name":{"kind":"Name","value":"email"}}]}}]}}]} as unknown as DocumentNode<SignUpMutation, SignUpMutationVariables>;
